+++ sphinx/ext/autodoc/directive.py
@@ -30,6 +30,8 @@ logger = logging.getLogger(__name__)
 AUTODOC_DEFAULT_OPTIONS = [
     'members',
     'undoc-members',
+    'no-index',
+    'no-index-entry',
     'inherited-members',
     'show-inheritance',
     'private-members',
@@ -40,7 +42,6 @@ AUTODOC_DEFAULT_OPTIONS = [
     'imported-members',
     'class-doc-from',
     'no-value',
-    'no-index-entry',
 ]
 
 AUTODOC_EXTENDABLE_OPTIONS = frozenset({